Testing module split
Separated the various tests into different submodules for easier test integration & reuse:
 * Regular testing definitions are now in `:testing:suite`
 * Custom test scripts are separate submodules in `:testing`, currently only `:testing:rdf-test-suite-js`
 * Moved `tests` SPARQL tests into `commonTest` for SPARQL
 * Moved benchmark related tests into `benchmarking` submodule, including necessary project modules for end-to-end benchmarking as a standalone build
Removed benchmarking git submodule
