+++ benchmarking/src/commonMain/kotlin/Main.kt
@@ -1,13 +1,15 @@
 import dev.tesserakt.util.printerrln
 import sparql.tests.compareIncrementalBasicGraphPatternOutput
+import sparql.tests.compareIncrementalSelectOutput
 
 suspend fun run(args: Array<String>) {
     when (args.size) {
         0 -> {
             println("Running built-in tests")
-            compareIncrementalBasicGraphPatternOutput()
-                .run()
-                .report()
+            val one = compareIncrementalSelectOutput().run()
+            val two = compareIncrementalBasicGraphPatternOutput().run()
+            one.report()
+            two.report()
         }
         2 -> {
             val (dataset, querypath) = args
