+++ benchmarking/src/commonMain/kotlin/sparql/Util.kt
