+++ benchmarking/src/jsMain/kotlin/comunica/Comunica.kt
