+++ benchmarking/src/jsMain/kotlin/comunica/Util.kt
