+++ benchmarking/src/jsMain/kotlin/sparql/Util.kt
