+++ benchmarking/src/jvmMain/kotlin/sparql/Util.kt
