+++ sparql/src/commonTest/kotlin/IncrementalQueryTest.kt
@@ -1,7 +1,3 @@
-package sparql
-
-import FOAF
-import createTestStore
 import dev.tesserakt.rdf.dsl.RdfContext.Companion.buildStore
 import dev.tesserakt.rdf.literalTerm
 import dev.tesserakt.rdf.namedTerm
