+++ sparql/src/commonTest/kotlin/Util.kt
