+++ testing/rdf-test-suite-js/src/jsMain/kotlin/Util.kt
